changed double to float for getProb, added new methods to Axe and Weapon, bug fix in Character.java to return this.weapon instead of return weapon

+++ src/main/java/commands/Throw.java
@@ -16,7 +16,7 @@ public class Throw extends Action {
         if (this.user.getWeapon() != null){
 
             // Get the weapon's throw hit probability and generate a random number to determine if it was a hit
-            Double throwProb = this.user.getWeapon().getProb();
+            float throwProb = this.user.getWeapon().getProb();
             Random rand = new Random();
             int hitNum = rand.nextInt(101);
